+++ src/hr/avrbanac/openglplayground/entities/Camera.java
@@ -1,6 +1,10 @@
 
 package hr.avrbanac.openglplayground.entities;
 
+import static hr.avrbanac.openglplayground.Globals.*;
+import hr.avrbanac.openglplayground.inputs.MouseButtonHandler;
+import hr.avrbanac.openglplayground.inputs.MousePosHandler;
+import hr.avrbanac.openglplayground.inputs.MouseScrollHandler;
 import hr.avrbanac.openglplayground.maths.Vector3f;
 import static org.lwjgl.glfw.GLFW.*;
 
@@ -8,18 +12,27 @@ import static org.lwjgl.glfw.GLFW.*;
  * Camera representation class.
  * 
  * @author avrbanac
- * @version 1.0.6
+ * @version 1.0.7
  */
 public class Camera {
     
-    private Vector3f position =  new Vector3f(0,0,0);
+    private Vector3f position;
     private float pitch;
     private float yaw;
     private float roll;
-    private long parentWindowID;
     
-    public Camera(long parentWindowID) {
-        this.parentWindowID = parentWindowID;
+    private Player player;
+    private float distanceFromPlayer;
+    private float angleAroundPlayer;
+    
+    public Camera(Player player) {
+        this.player = player;
+        position    =  new Vector3f(0,0,0);
+        pitch   = 20;
+        yaw     = 0;
+        roll    = 0;
+        distanceFromPlayer  = 50;
+        angleAroundPlayer   = 0;
     }
 
     public Vector3f getPosition() {
@@ -39,15 +52,13 @@ public class Camera {
     }
     
     public void move() {
-        if(glfwGetKey(parentWindowID, GLFW_KEY_KP_8) == GLFW_PRESS) position.z-=0.2f;
-        if(glfwGetKey(parentWindowID, GLFW_KEY_KP_2) == GLFW_PRESS) position.z+=0.2f;
-        if(glfwGetKey(parentWindowID, GLFW_KEY_KP_6) == GLFW_PRESS) position.x+=0.2f;
-        if(glfwGetKey(parentWindowID, GLFW_KEY_KP_4) == GLFW_PRESS) position.x-=0.2f;
-        if(glfwGetKey(parentWindowID, GLFW_KEY_KP_7) == GLFW_PRESS) position.y+=0.2f;
-        if(glfwGetKey(parentWindowID, GLFW_KEY_KP_1) == GLFW_PRESS) position.y-=0.2f;
-        if(glfwGetKey(parentWindowID, GLFW_KEY_KP_9) == GLFW_PRESS) pitch+=0.2f;
-        if(glfwGetKey(parentWindowID, GLFW_KEY_KP_3) == GLFW_PRESS) pitch-=0.2f;
+        calculateZoom();
+        calculatePitch();
+        calculateAngleAroundPlayer();
         
+        calculateCameraPosition(calculateHorizontalDistance(), calculateVerticalDistance());
+        
+        this.yaw = 180 - (player.getRotY() + angleAroundPlayer);
     }
     
     public void setPosition(Vector3f position) {
@@ -56,6 +67,51 @@ public class Camera {
     
     public void setPitch(float pitch) {
         this.pitch = pitch;
+    }
+    
+    private void calculateCameraPosition(float horizontalDistance, float verticalDistance) {
+        float theta     = player.getRotY() + angleAroundPlayer;
+        float offsetX   = (float) (horizontalDistance * Math.sin(Math.toRadians(theta)));
+        float offsetZ   = (float) (horizontalDistance * Math.cos(Math.toRadians(theta)));
+        
+        position.x      = player.getPosition().x - offsetX;
+        position.y      = player.getPosition().y + verticalDistance;
+        position.z      = player.getPosition().z - offsetZ;
+    }
+    
+    private void calculateZoom() {
+        float zoomLevel = MouseScrollHandler.getScrollY() * MOUSE_SENS_SCROLL;
+        distanceFromPlayer -= zoomLevel;
+        if (distanceFromPlayer < CAMERA_MIN_DIST) distanceFromPlayer = CAMERA_MIN_DIST;
+        if (distanceFromPlayer > CAMERA_MAX_DIST) distanceFromPlayer = CAMERA_MAX_DIST;
+    }
+    
+    private void calculatePitch() {
+        if(MouseButtonHandler.isButtonDown(GLFW_MOUSE_BUTTON_2)) {
+            if (MouseButtonHandler.hasButtonChangedState(GLFW_MOUSE_BUTTON_2)) MousePosHandler.resetDPos();
+            float pitchChange = MousePosHandler.getDY() * MOUSE_SENS_MOVE_Y;
+            pitch -= pitchChange;
+            if (pitch < CAMERA_MIN_PITCH) pitch = CAMERA_MIN_PITCH;
+            if (pitch > CAMERA_MAX_PITCH) pitch = CAMERA_MAX_PITCH;
+        }
+    }
+    
+    private void calculateAngleAroundPlayer() {
+        if(MouseButtonHandler.isButtonDown(GLFW_MOUSE_BUTTON_2)) {
+            if (MouseButtonHandler.hasButtonChangedState(GLFW_MOUSE_BUTTON_2)) MousePosHandler.resetDPos();
+            int dx = MousePosHandler.getDX();
+            float angleChange = dx * MOUSE_SENS_MOVE_X;
+            angleAroundPlayer -= angleChange;
+            //System.out.println("DX=" + String.valueOf(dx) + " CHANGE="+ String.valueOf(angleChange) + " ANGLE=" + String.valueOf(angleAroundPlayer));
+        }
+    }
+    
+    private float calculateHorizontalDistance() {
+        return (float) (distanceFromPlayer * Math.cos(Math.toRadians(pitch)));
+    }
+    
+    private float calculateVerticalDistance() {
+        return (float) (distanceFromPlayer * Math.sin(Math.toRadians(pitch)));
     }
     
 }